# add gabriel weights and more performant voronoi generation