--- a/libpysal/weights/__init__.py
+++ b/libpysal/weights/__init__.py
@@ -7,3 +7,4 @@ from .user import *
 from .set_operations import *
 from .spatial_lag import *
 from .raster import *
+from .gabriel import Gabriel, Delaunay, Relative_Neighborhood