--- a/libpysal/weights/tests/test_gabriel.py
+++ b/libpysal/weights/tests/test_gabriel.py
@@ -8,7 +8,7 @@ geoms = df.geometry.centroid
 coords = numpy.column_stack((geoms.x, geoms.y))
 
 
-def test_delaunay(coords):
+def test_delaunay():
     a = gabriel.Delaunay(coords)
     b = gabriel.Delaunay.from_dataframe(df)
 
@@ -17,7 +17,7 @@ def test_delaunay(coords):
     assert a[13] == {6: 1, 11: 1, 12: 1, 18: 1, 20: 1}
 
 
-def test_gabriel(coords):
+def test_gabriel():
     c = gabriel.Gabriel(coords)
     d = gabriel.Gabriel.from_dataframe(df)
     c2 = gabriel.Delaunay(coords)