This adds a revised `numba`-accelerated version of the Voronoi constructor (which avoids building the Voronoi polygons explicitly and instead opts to work directly with the Delaunay graph) and implements a Gabriel graph constructor (*edit: and a relative neighborhood constructor*) based on the new Voronoi/Delaunay constructor. The Gabriel graph is a subset of the Delaunay triangulation where the circle whose diameter is edge *ij* contains no other point *k*.

### The math for the Gabriel graph

Given the fast `QHull` Delaunay constructor, we can just start from the Delaunay triangulation and remove links *ij* if *k*, some other neighbor of *i*, is within the circle whose diameter is *ij*. 

To do this, we first can say (by [Thales' Theorem](https://en.wikipedia.org/wiki/Thales%27s_theorem)) that any triangle *ijk* with a side *ij* that is a circle diameter will be a *right* triangle when the non-diameter point *k* is *also* on the circle. This means that when *k* is *on the boundary* of the circle with *ij* as a diameter, then we have the Pythagorean relationship, *dij^2 = dik^2 + djk^2*. However, if *dij^2* is ever bigger than *dik^2 + djk^2*, then we know that triangle *ijk* is obtuse, and that *k* is "within" the circle whose diameter is *ij*. So, for focal *i* and each pair of neighbors *(j*,*k)*, prune Delaunay edge *ij* when *dij^2 > dik^2 + djk^2*. 

This needs:
- [ ] tests
- [x] comparison with R's [gabrielgraph](https://r-spatial.github.io/spdep/reference/graphneigh.html)
- [x] **decision**: if we use this graph constructor for Delaunay/Voronoi, it comes as about an order of magnitude speedup. But, it (a) cannot handle "clipping" the voronoi polygons to an arbitrary study extent (or, indeed, to a convex or alpha-hull) and (b) will then never be exactly equivalent to the Rook graph of a clipped Voronoi Diagram. But, this has the conceptual simplicity that it *only* converts the Delaunay triangulation into a graph, rather than building the voronoi cells, then clipping them, then constructing that Rook graph. So, input is needed on which we should use @sjsrey @martinfleis 
```
        Old Voronoi: 0.83s
        Raw Gabriel: 1.37s
        Compile: 1.88s
        New Delaunay: 0.04s
        Ref Gabriel: 0.07s
```