@ljwolf 2021-10-19
Results replicate with `spdep::gabrielgraph`!