@martinfleis 2021-11-05
> cannot handle "clipping" the voronoi polygons to an arbitrary study extent

I had the same issue with scipy.spatial.Voronoi in momepy and went around it by adding external points to the input. `momepy.Tessellation`, which is in its purest form a Voronoi tessellation based on points generated along the polygon boundary, requires a `limit` polygon, specifying the external limit of tessellation (see http://docs.momepy.org/en/stable/user_guide/elements/tessellation.html). We then take a convex hull of the limit, buffer it by twice the size of bounds and add points along the boundary of this buffer to the mix along those coming from buildings. That way we ensure that all tessellation cells always reach the limit unaffected by these external points and none of them ends in infinity as cells leading to infinity are only those based on the buffered hull. Those are dropped anyway (and remaining are clipped to `limit`).