@ljwolf 2021-11-09
> > cannot handle "clipping" the voronoi polygons to an arbitrary study extent
> 
> I had the same issue with scipy.spatial.Voronoi in momepy and went around it ...

yeah, in [the current implementation](https://github.com/pysal/libpysal/blob/5b386cfd182f131c174ef10f789e9c49acc7fd4d/libpysal/weights/contiguity.py#L480), we do something similar, allowing an automated clip to the bounding box, the convex hull, or the concave hull... But, this requires us to build the faces as polygons, which is generally slower than iterating over the tessellation in `numba`... 

But, there are quite a few cases where a clipped Voronoi diagram is not the same as a non-clipped one, thus the "Delaunay-like" triangulation implied by `Voronoi()` is not actually the same as the Rook graph of the clipped voronoi polygons. 

I think what we may do is:
1. keep the `Voronoi` around,
2. put it in the same module as this Delaunay & Mutual Neighbor stuff
3. write a vignette explaining why these aren't the same, and refer users to that vignette for the difference.