@martinfleis 2021-11-11
Ah, sure, that won't help Delaunay.

Since we do `voronoi_frames` + `Queen` under the hood now, I think we can deprecate `weights.Voronoi`. Users can do it themselves if they need a clipped version.

I am just not sure if I would use Delaunay constructor myself, knowing that it can result in some super weird neighbours, like two opposite corners of an array of points. Everything on the edges will be a bit sketchy.