@ljwolf 2021-11-18
> I am just not sure if I would use `Delaunay` constructor myself, knowing that it can result in some super weird neighbors

Yeah, definitely. But, my worry is that, in general, we can't get the gabriel graph from the clipped voronoi queen graph, since it's only a subset of the delaunay graph. 

Mathematically, I think there's a relationship between our "concave hull" alpha shape, the delaunay graph, and the gabriel graph, but I'm not sure how to characterize it off the top of my head. I think we need a proof that the radius defining the limiting concave hull is always larger than the largest gabriel neighbor radius. 

So, I put together a [small gist](https://gist.github.com/ljwolf/d7be4aeed8b0aba79b7dc9c1ec905035) to look into it. In this example, I simulate a banana-shaped pattern (where clipped Voronoi will be very different from the Delaunay triangulation), and check whether the Gabriel graph is a subgraph of the Queen Alpha-Hull clipped Voronoi.

![download](https://user-images.githubusercontent.com/2250995/142418136-7c53546a-2163-449e-bf99-be1a5a33d83e.png)

When I simulate 1000 random patterns with this shape, the gabriel is always a subset of the clipped Queen graph.