@ljwolf 2021-11-18
I should be clear: I'm in favor of keeping the Voronoi constructor around, but maybe "hiding" the raw delaunay one, or making `libpysal.weights.Delaunay` another name for `libpysal.weights.Voronoi` with `clip=None` as default.