@martinfleis 2021-11-18
Ahh, I should have paid more attention to the Gabriel graph. So the idea is to use `weights.Gabriel` where you now use clipped `weights.Voronoi`? That sounds good to me.

I wouldn't even hide `Delanuay`, just properly document it so users are aware of potential consequences.