@ljwolf 2021-12-10
This has tests and is passing. It only needs docstrings now and maybe a numba warning if computed w/o acceleration