@ljwolf 2021-12-11
@martinfleis, what would you think about exposing the momepy restricted voronoi tesselation in this PR, too, with a simple try/except wrapper class? 

I think it'd help for discoverability/completeness to expose it at the `weights` location, as we also should with the perimeter-weighted rook...