@martinfleis 2021-12-11
> @martinfleis, what would you think about exposing the momepy restricted voronoi tesselation in this PR, too, with a simple try/except wrapper class?

How would you do that? Create a tesselation GeoDataFrame and do Queen/Rook on top of that as we do in existing `Voronoi`? In principle that is why I guess, just people may be surprised that it is way slower than the other methods. And I have to check how does it behave when you pass point geometry type, that is not expected now.