@ljwolf 2021-12-11
Yes, that's exactly what I was thinking, under a `Voronoi_Tesselation` class... But, on reflection, I will stub this for a next release.