@martinfleis 2021-12-11
Thinking about it a bit more, it may be better to leave `momepy.Tessellation` out of this for now. It currently expects projected buildings on input (in meters) and we may want to try to generalise it a bit before plugging it in libpysal.