@ljwolf 2021-12-11
Yep, I totally agree. I think it could be generalized, but I don't want that to get in the way of releasing this... so merge here & then I'll cut releases of this + esda this weekend!