@ljwolf 2021-12-11
OK, all should be addressed plus a new test for the Relative Neighborhood Graph. I've also forced that to use binary weights (instead of distances) so that it behaves like the rest of the graphs in the module.