@martinfleis 2021-12-11
One more thing - we need to add it to docs (`docs/api.rst`).